Refactor padding retrieval for workspace layout

+++ library/hook/src/main/java/com/sevtinge/hyperceiler/hook/module/rules/home/layout/LayoutRules.kt
@@ -230,21 +230,21 @@ object LayoutRules : HomeBaseHookNew() {
 
             val sWorkspacePaddingTop = if (isSetWSPaddingTopHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_top", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_top", 0))
             } else {
                 -1
             }
 
             val sWorkspacePaddingBottom = if (isSetWSPaddingBottomHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_bottom", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_bottom", 0))
             } else {
                 -1
             }
 
             val sWorkspaceCellSide = if (isSetWSPaddingSideHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_horizontal", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_horizontal", 0))
             } else {
                 -1
             }
@@ -338,21 +338,21 @@ object LayoutRules : HomeBaseHookNew() {
 
             val sWorkspacePaddingTop = if (isSetWSPaddingTopHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_top", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_top", 0))
             } else {
                 -1
             }
 
             val sWorkspacePaddingBottom = if (isSetWSPaddingBottomHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_bottom", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_bottom", 0))
             } else {
                 -1
             }
 
             val sWorkspaceCellSide = if (isSetWSPaddingSideHook) {
                 DisplayUtils.dp2px(
-                    mPrefsMap.getInt("home_layout_workspace_padding_horizontal", 0).toFloat())
+                    mPrefsMap.getInt("home_layout_workspace_padding_horizontal", 0))
             } else {
                 -1
             }
